refactor: drop Firebase naming from multiplayer session-code helpers
- Move lib/firebase/sessionCodes.js → lib/multiplayer/sessionCodes.js (and the mirrored test file); update the page.jsx import
- Replace the "no Firebase deps" comment in the session-code helper with a provider-agnostic note
- Update the multiplayer_sessions migration's header comment to describe the table without naming Firebase as the prior storage
- Historical records (log.jsonl files, team-taboo meta.json improvement description) and Supabase's own config.toml third-party-auth template are left untouched
