+++ lib/multiplayer/sessionCodes.js
@@ -1,4 +1,4 @@
-// Session-code helpers. Pure functions — no Firebase deps — so they can be
+// Session-code helpers. Pure functions — no runtime deps — so they can be
 // used from Node.js scripts, Next.js pages, and unit tests.
 
 // Avoids visually ambiguous characters (0/O, 1/I, etc.) so players can copy
